[FRONT] WIP: Get plex servers
