+++ client/src/modules/settings/settings-configurations/SettingsConfigurations.js
@@ -1,12 +1,15 @@
 import React from 'react';
-import {PlexConfig} from "./components/PlexConfig";
+import {PlexConfig} from "./components/plex-config/PlexConfig";
+import {PlexConfigContextProvider} from "../../../contexts/PlexConfigContext";
 
 const SettingsConfigurations = (props) => {
   return (
     <div className="SettingsConfigurations container" data-testid="SettingsConfigurations">
       <h1 className="title is-1">Plex</h1>
       <hr/>
-      <PlexConfig location={props.location} />
+      <PlexConfigContextProvider>
+        <PlexConfig location={props.location} />
+      </PlexConfigContextProvider>
     </div>
   )
 };
